refactor: move auth credential and login owners
