+++ packages/shared-server/mod.ts
@@ -66,9 +66,10 @@ export type {
 export * from './rallar-system/client-state/snapshot/client-state-snapshot-read-through-cache.ts';
 export * from './rallar-system/client-state/snapshot/cached-client-state-service.ts';
 export * from './rallar-system/services/cached-group-state-service.ts';
-export * from './rallar-system/services/auth-login-service.ts';
+export * from './rallar-system/auth/login/authenticate-auth-user.ts';
+export * from './rallar-system/auth/login/prepare-auth-user-registration.ts';
 export * from './rallar-system/services/auth-state-mutations.ts';
-export * from './rallar-system/services/auth-credential-issuer.ts';
+export * from './rallar-system/auth/credentials/auth-credential-issuer.ts';
 export * from './rallar-system/services/AppAuthInboxService.ts';
 // prettier-ignore
 export {
