+++ packages/shared-server/rallar-system/auth/sessions/auth-session-proof-secret.ts
@@ -1,0 +1,13 @@
+import { hashAuthSecret } from '../credentials/hash-auth-secret.ts';
+import type {
+  IssuedAuthSession,
+  PersistedAuthSession,
+} from '../../repositories/AuthSessionRepository.ts';
+
+export async function authSessionProofSecret(
+  session: IssuedAuthSession | PersistedAuthSession,
+): Promise<string> {
+  return 'accessTokenDigest' in session
+    ? session.accessTokenDigest
+    : await hashAuthSecret(session.accessToken);
+}
