+++ packages/shared-server/rallar-system/repositories/AuthSessionRepository.ts
@@ -28,7 +28,7 @@ export {
     AUTH_LEGACY_PLAINTEXT_COMPATIBILITY_DEADLINE_EPOCH_MS,
     AUTH_LEGACY_PLAINTEXT_SCAN_LIMIT,
 } from './auth-legacy-compatibility.ts';
-export { hashAuthSecret } from './auth-secret-digest.ts';
+export { hashAuthSecret } from '../auth/credentials/hash-auth-secret.ts';
 export type {
     IssuedAgentSessionTicket,
     IssuedAuthSession,
