--- packages/shared-server/rallar-system/repositories/auth-secret-digest.ts
@@ -1,16 +1,0 @@
-export async function hashAuthSecret(secret: string): Promise<string> {
-    const digest = await crypto.subtle.digest(
-        'SHA-256',
-        new TextEncoder().encode(secret),
-    );
-    return toBase64Url(new Uint8Array(digest));
-}
-
-function toBase64Url(bytes: Uint8Array): string {
-    let binary = '';
-    for (const byte of bytes) binary += String.fromCharCode(byte);
-    return btoa(binary)
-        .replaceAll('+', '-')
-        .replaceAll('/', '_')
-        .replace(/=+$/u, '');
-}
