--- packages/shared-server/rallar-system/services/auth-session-proof-secret.ts
@@ -1,13 +1,0 @@
-import { hashAuthSecret } from '../repositories/AuthSessionRepository.ts';
-import type {
-    IssuedAuthSession,
-    PersistedAuthSession,
-} from '../repositories/AuthSessionRepository.ts';
-
-export async function authSessionProofSecret(
-    session: IssuedAuthSession | PersistedAuthSession,
-): Promise<string> {
-    return 'accessTokenDigest' in session
-        ? session.accessTokenDigest
-        : await hashAuthSecret(session.accessToken);
-}
